// apps/sartor-web/src/modules/health/core/enums/workout-type.enum.ts
export enum WorkoutType {
  RUNNING = "running",

  WALKING = "walking",

  HIKING = "hiking",

  CYCLING = "cycling",

  MOUNTAIN_BIKING = "mountain_biking",

  SWIMMING = "swimming",

  OPEN_WATER_SWIMMING = "open_water_swimming",

  STRENGTH = "strength",

  BODYBUILDING = "bodybuilding",

  POWERLIFTING = "powerlifting",

  CROSSFIT = "crossfit",

  HIIT = "hiit",

  CIRCUIT_TRAINING = "circuit_training",

  FUNCTIONAL_TRAINING = "functional_training",

  CALISTHENICS = "calisthenics",

  YOGA = "yoga",

  PILATES = "pilates",

  MOBILITY = "mobility",

  STRETCHING = "stretching",

  MEDITATION = "meditation",

  BOXING = "boxing",

  MMA = "mma",

  MARTIAL_ARTS = "martial_arts",

  DANCE = "dance",

  FOOTBALL = "football",

  BASKETBALL = "basketball",

  TENNIS = "tennis",

  PADEL = "padel",

  BADMINTON = "badminton",

  VOLLEYBALL = "volleyball",

  RUGBY = "rugby",

  SKI = "ski",

  SNOWBOARD = "snowboard",

  ROWING = "rowing",

  ELLIPTICAL = "elliptical",

  STAIR_CLIMBER = "stair_climber",

  JUMP_ROPE = "jump_rope",

  SURFING = "surfing",

  KAYAK = "kayak",

  CLIMBING = "climbing",

  HORSE_RIDING = "horse_riding",

  WORKOUT = "workout",

  OTHER = "other",
}